fix(ws-3): address review feedback
